LSPlayerCharacter 수정
1. Tag 추가
2. WBP_GameOver 생성 함수 PlayerCharacter 에서 PlayerController로 이동
3. CurrentDamage()에서 IsNearlyZero 를 0.f 로 변경
4. int32형 -> float로 변경

+++ Source/LSProject/Public/Character/LSPlayerCharacter.h
@@ -57,7 +57,7 @@ public:
 	float GetMaxHealth() const;
 
 	UFUNCTION(BlueprintCallable, Category = "UI")
-	void UpdateHealthBar(int32 Current, int32 Max);
+	void UpdateHealthBar(float Current, float Max);
 
 	UFUNCTION(BlueprintCallable, Category = "State")
 	void Equip();
